+++ community/src/main/java/kakaotech/community/domain/user/controller/UserController.java
@@ -1,11 +1,10 @@
 package kakaotech.community.domain.user.controller;
 
 import jakarta.validation.Valid;
-import kakaotech.community.domain.common.validator.Password;
 import kakaotech.community.domain.user.dto.UserRequest;
 import kakaotech.community.domain.user.dto.UserResponse;
 import kakaotech.community.domain.user.port.Token;
-import kakaotech.community.domain.user.service.AuthService;
+import kakaotech.community.domain.auth.service.AuthService;
 import kakaotech.community.domain.user.service.UserService;
 import kakaotech.community.global.apidoc.UserApiDocs;
 import kakaotech.community.global.auth.annotation.Authenticated;
@@ -25,13 +24,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/users")
 @RequiredArgsConstructor
 public class UserController implements UserApiDocs {
-    private final AuthService authService;
     private final UserService userService;
-
-    @PostMapping("/login")
-    public ResponseEntity<Token> login(@Valid @RequestBody UserRequest.Login request) {
-        return ResponseEntity.ok(authService.login(request.email(), request.password()));
-    }
 
     @PostMapping
     public ResponseEntity<UserResponse.Join> join(@Valid @ModelAttribute UserRequest.Join request) {
